+++ src/test/java/IDMaker/idmaker/ApplicationTests.java
@@ -1,3 +1,5 @@
+package IDMaker.idmaker;
+
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.boot.test.context.SpringBootTest;
