+++ src/test/java/IDMaker/idmaker/IDMakerTest.java
@@ -1,4 +1,4 @@
-package IDMaker.project.test;
+package IDMaker.idmaker;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -12,10 +12,8 @@ import java.util.stream.IntStream;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
 
-import IDMaker.idmaker.GenerationType;
 import IDMaker.idmaker.fixture.entity.TestClass;
 import IDMaker.idmaker.fixture.service.TestClassService;
-import IDMaker.idmaker.IDMakerUtils;
 
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
